Fix calculating pixel colour and painting the fractal + method header refactoring

+++ newton-fractal/AppWindow.hpp
@@ -1,3 +1,3 @@
 #pragma once
 
-auto drawWindow(int n, double WINDOW_WIDTH, double WINDOW_HEIGHT) -> void;
+auto drawWindow(int n, int WINDOW_WIDTH, int WINDOW_HEIGHT) -> void;
